requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.1 